refactor: move user repository classes to `domain/user` package
