+++ src/main/java/com/depromeet/archive/domain/user/UserAuthService.java
@@ -9,8 +9,6 @@ import com.depromeet.archive.exception.common.ResourceNotFoundException;
 import com.depromeet.archive.exception.user.LoginFailException;
 import com.depromeet.archive.exception.user.OAuthUserHasNotPasswordException;
 import com.depromeet.archive.infra.mail.MailService;
-import com.depromeet.archive.infra.user.jpa.PasswordUserRepository;
-import com.depromeet.archive.infra.user.jpa.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
